+++ R/zzz.R
@@ -1,6 +1,12 @@
 .onLoad <- function(libname, pkgname) {
   ns <- asNamespace(pkgname)
   delayedAssign("GITIGNORED", gitignore_items(), eval.env = ns, assign.env = ns)
-  renv_has_consent()
+  # Check for implicit {renv} consent. If the user has used it before, we should
+  # use it in the {sandpaper} lesson, unless the user has explicitly told us not
+  # to.
+  op <- getOption("sandpaper.use_renv")
+  if (is.null(op)) {
+    renv_has_consent()
+  }
   invisible()
 }